+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [master](https://github.com/anmonteiro/lumo/compare/1.3.0...HEAD) (unreleased)
 
+### New features
+
+- Return exit code `1` on exception ([#58](https://github.com/anmonteiro/lumo/issues/58)).
+
 ## [1.3.0](https://github.com/anmonteiro/lumo/compare/1.2.0...1.3.0) (2017-04-02)
 
 ### New features